refactor: sonarlint fixes
Found a bug in getChannelGuid that tried to compare a String with an
Integer that would always result in false. There doesn't seem to be a
test for it though since it was never caught, although
AssetServiceBeanIntTest.collectAssetMTRequestIridiumTransponder failed
before adding a dnid/memberNumber null check.
Might not be used since no one has complained (?) about not enriching
with channels?