test: removes user transaction (TransactionalTests) from tests
Having each test case wrapped in a `begin commit` UserTransaction
doesn't do the same thing as the default container transaction does
which meant that some/many test cases tested a setup that doesn't occur.
Most notably, there are test cases that should hit the DB (commit) and
fail but doesn't as the commit is done after the test case, or after the
method that should've thrown the exception, is called.

Using the default container transaction setup better tests these
scenarios.

Removed cascade ALL on @ManyToOne for PollBase (MobileTerminal) and
MobileTerminal (MobileTerminalPlugin) to avoid
PersistentObjectException. It's unclear if the persist behaviour is only
a problem in tests or in practice.

Fixes naming (ending with IT) for certain test classes as they were not
picked up and run during the normal test phase.